@@ -1,19 +1,23 @@
 # SPDX-FileCopyrightText: 2025 GitHub
 # SPDX-License-Identifier: MIT
 
-from enum import Enum
-import logging
 import importlib.resources
+from enum import Enum
+
 import yaml
+
 
 class BadToolNameError(Exception):
     pass
 
+
 class VersionException(Exception):
     pass
 
+
 class FileTypeException(Exception):
     pass
+
 
 class AvailableToolType(Enum):
     Personality = "personality"
@@ -22,11 +26,13 @@ class AvailableToolType(Enum):
     Toolbox = "toolbox"
     ModelConfig = "model_config"
 
+
 class AvailableTools:
     """
     This class is used for storing dictionaries of all the available
     personalities, taskflows, and prompts.
     """
+
     def __init__(self):
         self.__yamlcache = {}
 
@@ -56,15 +62,17 @@ class AvailableTools:
         except KeyError:
             pass
         # Split the string to get the package and filename.
-        components = toolname.rsplit('.', 1)
+        components = toolname.rsplit(".", 1)
         if len(components) != 2:
-            raise BadToolNameError(f'Not a valid toolname: "{toolname}". It should be something like: "packagename.filename"')
+            raise BadToolNameError(
+                f'Not a valid toolname: "{toolname}". It should be something like: "packagename.filename"'
+            )
         package = components[0]
         filename = components[1]
         try:
             d = importlib.resources.files(package)
             if not d.is_dir():
-                raise BadToolNameError(f'Cannot load {toolname} because {d} is not a valid directory.')
+                raise BadToolNameError(f"Cannot load {toolname} because {d} is not a valid directory.")
             f = d.joinpath(filename + ".yaml")
             with open(f) as s:
                 y = yaml.safe_load(s)
@@ -89,16 +97,15 @@ class AvailableTools:
                     )
                 filetype = header['filetype'] 
                 if filetype != tooltype.value:
-                    raise FileTypeException(
-                        f'Error in {f}: expected filetype to be {tooltype}, but it\'s {filetype}.')
+                    raise FileTypeException(f"Error in {f}: expected filetype to be {tooltype}, but it's {filetype}.")
                 if tooltype not in self.__yamlcache:
                     self.__yamlcache[tooltype] = {}
                 self.__yamlcache[tooltype][toolname] = y
                 return y
         except ModuleNotFoundError as e:
-            raise BadToolNameError(f'Cannot load {toolname}: {e}')
+            raise BadToolNameError(f"Cannot load {toolname}: {e}")
         except FileNotFoundError:
             # deal with editor temp files etc. that might have disappeared
-            raise BadToolNameError(f'Cannot load {toolname} because {f} is not a valid file.')
+            raise BadToolNameError(f"Cannot load {toolname} because {f} is not a valid file.")
         except ValueError as e:
-            raise BadToolNameError(f'Cannot load {toolname}: {e}')
+            raise BadToolNameError(f"Cannot load {toolname}: {e}")
