@@ -3,7 +3,7 @@
 
 import os
 import jinja2
-
+import re
 
 def swap_env(s: str) -> str:
     """Replace {{ env('VAR') }} patterns in string with environment values.
@@ -36,6 +36,7 @@ def swap_env(s: str) -> str:
     except jinja2.TemplateError:
         # Not a template or failed to render, return as-is
         return s
+
 
 class TmpEnv:
     def __init__(self, env):
