@@ -1,32 +1,35 @@
 # SPDX-FileCopyrightText: 2025 GitHub
 # SPDX-License-Identifier: MIT
 
-import logging
-#from mcp.server.fastmcp import FastMCP
-from fastmcp import FastMCP # move to FastMCP 2.0
 import json
+import logging
 from pathlib import Path
-from seclab_taskflow_agent.path_utils import mcp_data_dir, log_file_name
+
+# from mcp.server.fastmcp import FastMCP
+from fastmcp import FastMCP  # move to FastMCP 2.0
+
+from seclab_taskflow_agent.path_utils import log_file_name, mcp_data_dir
 
 logging.basicConfig(
     level=logging.DEBUG,
-    format='%(asctime)s - %(levelname)s - %(message)s',
-    filename=log_file_name('mcp_logbook.log'),
-    filemode='a'
+    format="%(asctime)s - %(levelname)s - %(message)s",
+    filename=log_file_name("mcp_logbook.log"),
+    filemode="a",
 )
 
 mcp = FastMCP("Logbook")
 
 LOG = {}
 
-LOGBOOK = mcp_data_dir('seclab-taskflow-agent', 'logbook', 'LOGBOOK_STATE_DIR') / Path("logbook.json")
+LOGBOOK = mcp_data_dir("seclab-taskflow-agent", "logbook", "LOGBOOK_STATE_DIR") / Path("logbook.json")
+
 
 def ensure_log():
     global LOG
     global LOGBOOK
     try:
         LOGBOOK.parent.mkdir(exist_ok=True, parents=True)
-        with open(LOGBOOK, 'x') as logbook:
+        with open(LOGBOOK, "x") as logbook:
             logbook.write(json.dumps(LOG, indent=2))
             logbook.flush()
     except FileExistsError:
@@ -37,7 +40,7 @@ def deflate_log():
     ensure_log()
     global LOG
     global LOGBOOK
-    with open(LOGBOOK, 'w') as logbook:
+    with open(LOGBOOK, "w") as logbook:
         logbook.write(json.dumps(LOG, indent=2))
         logbook.flush()
 
@@ -46,7 +49,7 @@ def inflate_log():
     ensure_log()
     global LOG
     global LOGBOOK
-    with open(LOGBOOK, 'r') as logbook:
+    with open(LOGBOOK) as logbook:
         LOG = json.loads(logbook.read())
 
 
@@ -56,38 +59,45 @@ def with_log(f):
         ret = f(*args, **kwargs)
         deflate_log()
         return ret
+
     return wrapper
 
 
 @mcp.tool()
 def logbook_write(entry: str, key: str) -> str:
     """Appends a logbook entry to an identifying key. This lets you write to your logbook."""
+
     @with_log
     def _logbook_write(entry: str, key: str) -> str:
         global LOG
         LOG[key] = LOG.get(key, []) + [entry]
         return f"Stored logbook entry for `{key}`"
+
     return _logbook_write(entry, key)
 
 
 @mcp.tool()
 def logbook_read(key: str) -> str:
     """Reads the entries stored for an identifying key. This lets you read from your logbook."""
+
     @with_log
     def _logbook_read(key: str) -> str:
         global LOG
         return json.dumps(LOG.get(key, []), indent=2)
+
     return _logbook_read(key)
 
 
 @mcp.tool()
 def logbook_erase(key: str) -> str:
     """Erase the entries stored for an identifying key. This lets you erase in your logbook."""
+
     @with_log
     def _logbook_erase(key) -> str:
         global LOG
         LOG[key] = []
         return f"Erased logbook entries stored for `{key}`"
+
     return _logbook_erase(key)
 
 
