@@ -3,6 +3,7 @@
 
 from typing import Any
 
+
 class Backend:
     def __init__(self, memcache_state_dir: str):
         self.memcache_state_dir = memcache_state_dir
@@ -15,9 +16,9 @@ class Backend:
 
     def add_state(self, key: str, value: Any) -> str:
         pass
-    
+
     def list_keys(self) -> str:
-        pass    
+        pass
 
     def clear_cache(self) -> str:
         pass
