@@ -1,25 +1,27 @@
 # SPDX-FileCopyrightText: 2025 GitHub
 # SPDX-License-Identifier: MIT
 
+import json
 import os
 from pathlib import Path
+from typing import Any
+
 from sqlalchemy import create_engine
 from sqlalchemy.orm import Session
-from typing import Any
-import json
 
-from .sql_models import KeyValue, Base
 from .backend import Backend
+from .sql_models import Base, KeyValue
+
 
 class SqliteBackend(Backend):
     def __init__(self, memcache_state_dir: str):
         super().__init__(memcache_state_dir)
         if not Path(self.memcache_state_dir).exists():
-            db_dir = 'sqlite://'
+            db_dir = "sqlite://"
         else:
-            db_dir = f'sqlite:///{os.path.abspath(self.memcache_state_dir)}/memory.db'
+            db_dir = f"sqlite:///{os.path.abspath(self.memcache_state_dir)}/memory.db"
         self.engine = create_engine(db_dir, echo=False)
-        Base.metadata.create_all(self.engine, tables = [KeyValue.__table__])
+        Base.metadata.create_all(self.engine, tables=[KeyValue.__table__])
 
     def set_state(self, key: str, value: Any) -> str:
         with Session(self.engine) as session:
@@ -28,7 +30,7 @@ class SqliteBackend(Backend):
             session.add(kv)
             session.commit()
         return 'f"Stored value in memory for `{key}`"'
-    
+
     def get_state(self, key: str) -> Any:
         with Session(self.engine) as session:
             values = session.query(KeyValue).filter_by(key=key).all()
@@ -44,14 +46,14 @@ class SqliteBackend(Backend):
             for r in results[1:]:
                 existing.append(r)
             return existing
-        elif hasattr(existing, '__add__'):
+        if hasattr(existing, "__add__"):
             try:
                 for r in results[1:]:
                     existing += r
                 return existing
             except TypeError:
                 return results
-    
+
     def add_state(self, key, value):
         with Session(self.engine) as session:
             kv = KeyValue(key=key, value=json.dumps(value))
@@ -64,8 +66,8 @@ class SqliteBackend(Backend):
             keys = session.query(KeyValue.key).distinct().all()
         content = ["IMPORTANT: your known memcache keys are now:\n"]
         content += [f"- {key[0]}" for key in keys]
-        return '\n'.join(content)        
-    
+        return "\n".join(content)
+
     def get_all_entries(self) -> str:
         with Session(self.engine) as session:
             entries = session.query(KeyValue).all()
@@ -77,15 +79,10 @@ class SqliteBackend(Backend):
             session.commit()
         if result:
             return f"Deleted key `{key}` from memory cache."
-        else:
-            return f"Key `{key}` not found in memory cache."
+        return f"Key `{key}` not found in memory cache."
 
     def clear_cache(self) -> str:
         with Session(self.engine) as session:
             session.query(KeyValue).delete()
             session.commit()
         return "Cleared all keys in memory cache."
-
-
-
-
