@@ -5,10 +5,13 @@
 Test API endpoint configuration.
 """
 
-import pytest
 import os
 from urllib.parse import urlparse
-from seclab_taskflow_agent.capi import get_AI_endpoint, AI_API_ENDPOINT_ENUM, list_capi_models
+
+import pytest
+
+from seclab_taskflow_agent.capi import AI_API_ENDPOINT_ENUM, get_AI_endpoint, list_capi_models
+
 
 class TestAPIEndpoint:
     """Test API endpoint configuration."""
@@ -18,7 +21,7 @@ class TestAPIEndpoint:
         # When no env var is set, it should default to models.github.ai/inference
         try:
             # Save original env
-            original_env = os.environ.pop('AI_API_ENDPOINT', None)
+            original_env = os.environ.pop("AI_API_ENDPOINT", None)
             endpoint = get_AI_endpoint()
             assert endpoint is not None
             assert isinstance(endpoint, str)
@@ -26,48 +29,49 @@ class TestAPIEndpoint:
         finally:
             # Restore original env
             if original_env:
-                os.environ['AI_API_ENDPOINT'] = original_env
+                os.environ["AI_API_ENDPOINT"] = original_env
 
     def test_api_endpoint_env_override(self):
         """Test that AI_API_ENDPOINT can be overridden by environment variable."""
         try:
             # Save original env
-            original_env = os.environ.pop('AI_API_ENDPOINT', None)
+            original_env = os.environ.pop("AI_API_ENDPOINT", None)
             # Set different endpoint
-            test_endpoint = 'https://api.githubcopilot.com'
-            os.environ['AI_API_ENDPOINT'] = test_endpoint
+            test_endpoint = "https://api.githubcopilot.com"
+            os.environ["AI_API_ENDPOINT"] = test_endpoint
 
             assert get_AI_endpoint() == test_endpoint
         finally:
             # Restore original env
             if original_env:
-                os.environ['AI_API_ENDPOINT'] = original_env
+                os.environ["AI_API_ENDPOINT"] = original_env
 
     def test_to_url_models_github(self):
         """Test to_url method for models.github.ai endpoint."""
         endpoint = AI_API_ENDPOINT_ENUM.AI_API_MODELS_GITHUB
-        assert endpoint.to_url() == 'https://models.github.ai/inference'
+        assert endpoint.to_url() == "https://models.github.ai/inference"
 
     def test_to_url_githubcopilot(self):
         """Test to_url method for GitHub Copilot endpoint."""
         endpoint = AI_API_ENDPOINT_ENUM.AI_API_GITHUBCOPILOT
-        assert endpoint.to_url() == 'https://api.githubcopilot.com'
+        assert endpoint.to_url() == "https://api.githubcopilot.com"
 
     def test_to_url_openai(self):
         """Test to_url method for OpenAI endpoint."""
         endpoint = AI_API_ENDPOINT_ENUM.AI_API_OPENAI
-        assert endpoint.to_url() == 'https://api.openai.com/v1'
+        assert endpoint.to_url() == "https://api.openai.com/v1"
 
     def test_unsupported_endpoint(self, monkeypatch):
         """Test that unsupported API endpoint raises ValueError."""
-        api_endpoint = 'https://unsupported.example.com'
-        monkeypatch.setenv('AI_API_ENDPOINT', api_endpoint)
+        api_endpoint = "https://unsupported.example.com"
+        monkeypatch.setenv("AI_API_ENDPOINT", api_endpoint)
         with pytest.raises(ValueError) as excinfo:
             list_capi_models("abc")
         msg = str(excinfo.value)
-        assert 'Unsupported Model Endpoint' in msg
-        assert 'https://models.github.ai/inference' in msg
-        assert 'https://api.githubcopilot.com' in msg
+        assert "Unsupported Model Endpoint" in msg
+        assert "https://models.github.ai/inference" in msg
+        assert "https://api.githubcopilot.com" in msg
 
-if __name__ == '__main__':
-    pytest.main([__file__, '-v'])
+
+if __name__ == "__main__":
+    pytest.main([__file__, "-v"])
